requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic>=2.7.0
pydantic-settings>=2.3.0
openai>=1.30.0
yt-dlp>=2024.5.1
paddlepaddle>=2.6.1
paddleocr>=2.7.3
Pillow>=10.3.0
aiofiles>=23.2.1
httpx>=0.27.0
python-dotenv>=1.0.1
orjson>=3.10.0
pymupdf>=1.24.0
websockets>=12.0
